Finish debug project1

+++ cs143b/src/main/java/data/PRManager.java
@@ -10,26 +10,55 @@ public class PRManager {
 	private ReadyList RL;
 	
 	public PRManager(){
-		this.runningProcess = new PCB("Init", 0);
-//		this.runningProcess.getStatus().type = PCB.Type.RUNNING;
-		System.out.println("Init is running");
-		allResources = new LinkedList<RCB>();
-		allResources.add(new RCB("R1", 1));
-		allResources.add(new RCB("R2", 1));
-		RL = new ReadyList();
-		this.RL.insert(this.runningProcess);
-		IO = new RCB("IO", 1);
+		init();
+		System.out.println("Process Init is running");
 	}
 	
 	public void init(){
 		this.runningProcess = new PCB("Init", 0);
-		System.out.println("Init is running");
+		this.runningProcess.getStatus().type = PCB.Type.RUNNING;
 		allResources = new LinkedList<RCB>();
 		allResources.add(new RCB("R1", 1));
 		allResources.add(new RCB("R2", 1));
+		allResources.add(new RCB("R3", 1));
+		allResources.add(new RCB("R4", 1));
 		RL = new ReadyList();
-		this.RL.insert(this.runningProcess);
 		IO = new RCB("IO", 1);
+	}
+	
+	public void execute(String line){
+		String[] tokens = line.split(" ");
+		String command = tokens[0];
+		if (command.equals("")) {
+			System.out.println();
+			return;
+		}
+		if (command.equals("init")) {
+			init();
+		} else if (command.equals("cr")) {
+			createProcess(tokens[1], Integer.parseInt(tokens[2]));
+		} else if(command.equals("de")){ 
+			destroyProcess(tokens[1]);
+		} else if(command.equals("to")){ 
+			timeOut();
+		} else if(command.equals("req")){
+			requestResource(tokens[1]);
+		} else if(command.equals("rel")){
+			releaseResource(tokens[1], true);
+		} else if(command.equals("rio")){
+			requestIO();
+		} else if(command.equals("ioc")){
+			IOCompletion();
+		} else if(command.equals("lsp")){
+			listAllProcessesAndStatus();
+		} else if(command.equals("lsr")){
+			listAllResourcesAndStatus();
+		} else if(command.equals("infop")){
+			printProcessInfo(tokens[1]);
+		} else if(command.equals("infor")){
+			printResourceInfo(tokens[1]);
+		}
+		printCurrentRunningProcess();
 	}
 	
 	public void createProcess(String pid, int priority){
@@ -45,12 +74,14 @@ public class PRManager {
 		
 		// 4. reschedule
 		this.scheduler();
-		
-//		System.out.println(RL.toString());
 	}
 	
 	public void destroyProcess(String pid){
-		PCB pcb = this.RL.get(pid);
+		PCB pcb = getPCB(pid);
+		if(pcb == null){
+			System.out.println("Process " + pid + " doesn't exist!");
+			return;
+		}
 		killTree(pcb);
 		this.scheduler();
 	}
@@ -62,9 +93,38 @@ public class PRManager {
 		
 		// free resources
 		for(RCB rcb : pcb.getResourceList())
-			releaseResource(rcb.getRid());
+			releaseResource(rcb.getRid(), false);
+		
 		// update pointer
-		this.RL.remove(pcb);
+		if(this.runningProcess != null && this.runningProcess.equals(pcb))
+			this.runningProcess = null;
+		else
+			this.RL.remove(pcb);
+		pcb = null;
+	}
+	
+	public PCB getPCB(String pid){
+		PCB pcb = null;
+		if(this.runningProcess.getPid().equals(pid))
+			pcb = this.runningProcess;
+		else{
+			pcb = this.RL.get(pid);
+		}
+		if(pcb == null)
+			pcb = getPCBFromWaitingList(pid);
+		return pcb;
+	}
+	
+	public PCB getPCBFromWaitingList(String pid){
+		for(RCB rcb : this.allResources){
+			Iterator<PCB> it = rcb.getWaitingList().iterator();
+			while(it.hasNext()){
+				PCB pcb = it.next();
+				if(pcb.getPid().equals(pid))
+					return pcb;
+			}
+		}
+		return null;
 	}
 	
 	public void requestResource(String rid){
@@ -79,13 +139,29 @@ public class PRManager {
 			rcb.getWaitingList().add(this.runningProcess);
 			scheduler();
 		}
-		
-//		System.out.println("Current ResourceList: " + this.allResources);
 	}
 	
-	public void releaseResource(String rid){
-		RCB rcb = this.runningProcess.getRCB(rid);
-		this.runningProcess.getResourceList().remove(rcb);
+	public void releaseResource(String rid, boolean releaseByRunningProcess){
+		RCB rcb = null;
+		if(releaseByRunningProcess){
+			rcb = this.runningProcess.getRCB(rid);
+			if(rcb == null){
+				System.out.println("Release " + rid + " failed, the current running process doesn't hold this resource!");
+				return;
+			}
+			this.runningProcess.getResourceList().remove(rcb);
+		} else{ // release when destroy the process
+			for(RCB resource : this.allResources){
+				if(resource.getRid().equals(rid)){
+					rcb = resource;
+					break;
+				}
+			}
+			if(rcb == null){
+				System.out.println("Release " + rid + " failed, this resource doesn't exist!");
+				return;
+			}
+		}
 		LinkedList<PCB> waitingList = rcb.getWaitingList();
 		if(waitingList.isEmpty())
 			rcb.release(1);
@@ -97,8 +173,6 @@ public class PRManager {
 			pcb.getResourceList().add(rcb);
 			scheduler();
 		}
-		
-//		System.out.println("Current ResourceList: " + this.allResources);
 	}
 	
 	public RCB getRCB(String rid){
@@ -129,16 +203,14 @@ public class PRManager {
 	
 	// function call simulate time out
 	public void timeOut(){
-		// 1. remove q from Ready List
-		this.RL.remove(this.runningProcess);
-		
-		// 2. insert q into Ready List
+		// 1. insert q into Ready List
 		this.RL.insert(this.runningProcess);
+		
+		// 2. change q.status to ready
+		this.runningProcess.getStatus().type = PCB.Type.READY;
 		
 		// 3. reschedule
 		this.scheduler();
-		
-		System.out.println(RL.toString());
 	}
 	
 	public void scheduler(){
@@ -146,9 +218,10 @@ public class PRManager {
 		PCB highest = RL.getPCBWithHighPriority();
 		
 		// 2. under some conditions, do context switch
-		if(this.runningProcess.getPriority() < highest.getPriority()	//(1) create/release
+		if(this.runningProcess == null	//(3) destroy
+				|| this.runningProcess.getPriority() < highest.getPriority()	//(1) create/release
 				|| this.runningProcess.getStatus().type != PCB.Type.RUNNING	//(2) request/time-out
-				|| this.runningProcess == null)	//(3) destroy
+				)	
 			preempt(highest);
 	}
 	
@@ -161,7 +234,6 @@ public class PRManager {
 		highest.getStatus().type = PCB.Type.RUNNING;
 		this.runningProcess = highest;
 		this.RL.remove(highest);
-//		System.out.println("Process " + highest.getPid() + " is running");
 	}
 	
 	public void printCurrentRunningProcess(){
@@ -175,14 +247,29 @@ public class PRManager {
 	
 	public void printWaitingList(){
 		System.out.println("The waiting list is :");
+		boolean flag = false;
 		for(RCB rcb : this.allResources){
-			if(!rcb.getWaitingList().isEmpty())
+			if(!rcb.getWaitingList().isEmpty()){
+				flag = true;
 				System.out.print(rcb.getWaitingList() + "	");
+			}
 		}
-		System.out.println();
+		if(!flag)
+			System.out.println();
 	}
 	
 	public void listAllResourcesAndStatus(){
-		
+		for(RCB rcb : this.allResources)
+			System.out.println(rcb.getInfo());
+	}
+	
+	public void printProcessInfo(String pid){
+		PCB pcb = getPCB(pid);
+		System.out.println(pcb.getInfo());
+	}
+	
+	public void printResourceInfo(String rid){
+		RCB rcb = this.getRCB(rid);
+		System.out.println(rcb.getInfo());
 	}
 }
